Prevent rounding issues by serialising site latitude and longitude as a float not a double

+++ api/src/main/java/au/org/aodn/nrmn/restapi/dto/site/SiteListItem.java
@@ -32,10 +32,10 @@ public class SiteListItem extends RepresentationModel<SiteListItem> {
     String country;
 
     @Schema(title = "Latitude")
-    Double latitude;
+    Float latitude;
 
     @Schema(title = "Longitude")
-    Double longitude;
+    Float longitude;
 
     @Schema(title = "Is Active")
     Boolean isActive;
